decompose_range_check: Move to submodule.
