+++ src/range_check/example3_broken/table.rs
@@ -70,7 +70,6 @@ impl<F: FieldExt, const NUM_BITS: usize, const RANGE: usize> RangeTableConfig<F,
                         )?;
                         offset += 1;
                     }
-
                 }
 
                 Ok(())
